Deleting the user_repository package and realocating your functions in todo_list_repository
