+++ repositories/__init__.py
@@ -1,2 +1,1 @@
-from .user_repository import UserRepository
 from .todo_list_repository import TodoListRepository
